// packages/domain-pack/src/types.ts
export type DiscoverMissingExpectation = {
  id: string;
  packExpectationId: string;
  requirementClass: PackExpectationRequirementClass;
  expectedDocumentType: string;
  familyRole: string;
  groupId: string;
  reasonExpected: string;
  sequenceOrder: number;
};